+++ src/shopping/core/RegularItem.java
@@ -2,18 +2,18 @@ package shopping.core;
 
 public class RegularItem extends CartItem {
 
-	public RegularItem(Product p, int qty) {
-		super(p, qty);
-	}
+  public RegularItem(Product p, int qty) {
+    super(p, qty);
+  }
 
-	@Override
-	double price() {
-		return unitPrice() * quantity;
-	}
+  @Override
+  int price() {
+    return unitPrice() * quantity;
+  }
 
-	@Override
-	protected String itemType() {
-		return "Regular Item";
-	}	
-	
+  @Override
+  protected String itemType() {
+    return "Regular Item";
+  }
+
 }
